requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Pydantic for data validation
pydantic==2.5.0

# PDF processing
PyPDF2==3.0.1

# Machine Learning and NLP
transformers==4.36.0
torch==2.1.0
tokenizers==0.15.0

# Text processing and fuzzy matching
rapidfuzz==3.5.2

# HTTP requests (if needed)
requests==2.31.0

# JSON handling (built-in, but listed for completeness)
# json - built-in module

# Regular expressions (built-in)
# re - built-in module

# Path handling (built-in)
# pathlib - built-in module

# Logging (built-in)
# logging - built-in module

# Optional: For enhanced text processing
nltk==3.8.1
spacy==3.7.2

# Optional: For database support if needed later
# sqlalchemy==2.0.23
# databases==0.8.0

# Optional: For caching
# redis==5.0.1

# Development dependencies (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1